Factor out State code
