+++ mlabs/mlabs-plutus-use-cases.cabal
@@ -43,6 +43,8 @@ library
   default-language: Haskell2010
   hs-source-dirs:      src/
   exposed-modules:
+        Mlabs.Control.Check
+        Mlabs.Control.Monad.State
         Mlabs.Data.AssocMap
         Mlabs.Data.List
         Mlabs.Data.Ord
