+++ mlabs/src/Mlabs/Lending/Logic/React.hs
@@ -21,6 +21,7 @@ import qualified PlutusTx.These as PlutusTx
 import Control.Monad.Except hiding (Functor(..), mapM)
 import Control.Monad.State.Strict hiding (Functor(..), mapM)
 
+import Mlabs.Control.Check
 import Mlabs.Lending.Logic.Emulator.Blockchain
 import Mlabs.Lending.Logic.InterestRate (addDeposit)
 import Mlabs.Lending.Logic.State
